Adding get identity API to edgelet client
