+++ mqtt/edgelet-client/src/workload/models/mod.rs
@@ -1,7 +1,11 @@
+mod cert_response;
+mod identity_cert;
 mod server_cert;
 mod sign_request;
 mod trust_bundle;
 
+pub use cert_response::*;
+pub use identity_cert::*;
 pub use server_cert::*;
 pub use sign_request::*;
 pub use trust_bundle::*;
