+++ scripts/linux/cross-platform-rust-build.sh
@@ -83,7 +83,7 @@ case "$PACKAGE_ARCH" in
         ;;
 
     'arm32v7')
-        RUST_TARGET='armv7-unknown-linux-gnueabihf'
+        RUST_TARGET='armv7-unknown-linux-musleabihf'
         ;;
 
     'aarch64')
@@ -153,31 +153,57 @@ case "$PACKAGE_OS.$PACKAGE_ARCH" in
         ;;
 
     ubuntu18.04.arm32v7)
+        # The below SETUP was copied from https://github.com/emk/rust-musl-builder/blob/master/Dockerfile.
         SETUP_COMMAND=$'
-            sources="$(cat /etc/apt/sources.list | grep -E \'^[^#]\')" &&
-            # Update existing repos to be specifically for amd64
-            echo "$sources" | sed -e \'s/^deb /deb [arch=amd64] /g\' > /etc/apt/sources.list &&
-            # Add armhf repos
-            echo "$sources" |
-                sed -e \'s/^deb /deb [arch=armhf] /g\' \
-                    -e \'s| http://archive.ubuntu.com/ubuntu/ | http://ports.ubuntu.com/ubuntu-ports/ |g\' \
-                    -e \'s| http://security.ubuntu.com/ubuntu/ | http://ports.ubuntu.com/ubuntu-ports/ |g\' \
-                    >> /etc/apt/sources.list &&
-
-            dpkg --add-architecture armhf &&
-            apt-get update &&
-            apt-get upgrade -y &&
-            apt-get install -y --no-install-recommends \
-                binutils build-essential ca-certificates cmake curl debhelper dh-systemd file git make \
-                gcc g++ \
-                gcc-arm-linux-gnueabihf g++-arm-linux-gnueabihf \
-                libcurl4-openssl-dev:armhf libssl-dev:armhf uuid-dev:armhf &&
-
-            mkdir -p ~/.cargo &&
-            echo \'[target.armv7-unknown-linux-gnueabihf]\' > ~/.cargo/config &&
-            echo \'linker = \"arm-linux-gnueabihf-gcc\"\' >> ~/.cargo/config &&
-            export ARMV7_UNKNOWN_LINUX_GNUEABIHF_OPENSSL_LIB_DIR=/usr/lib/arm-linux-gnueabihf &&
-            export ARMV7_UNKNOWN_LINUX_GNUEABIHF_OPENSSL_INCLUDE_DIR=/usr/include &&
+            OPENSSL_VERSION=1.1.1g
+            apt-get update && \
+            apt-get install -y \
+                build-essential \
+                cmake \
+                curl \
+                file \
+                git \
+                musl-dev \
+                musl-tools \
+                libpq-dev \
+                libsqlite-dev \
+                libssl-dev \
+                linux-libc-dev \
+                pkgconf \
+                sudo \
+                xutils-dev \
+                gcc-multilib-arm-linux-gnueabihf \
+                && \
+            apt-get clean && rm -rf /var/lib/apt/lists/* && \
+            useradd rust --user-group --create-home --shell /bin/bash --groups sudo && \
+            echo "Building OpenSSL" && \
+            ls /usr/include/linux && \
+            sudo mkdir -p /usr/local/musl/include && \
+            sudo ln -s /usr/include/linux /usr/local/musl/include/linux && \
+            sudo ln -s /usr/include/x86_64-linux-gnu/asm /usr/local/musl/include/asm && \
+            sudo ln -s /usr/include/asm-generic /usr/local/musl/include/asm-generic && \
+            cd /tmp && \
+            short_version="$(echo "$OPENSSL_VERSION" | sed s\'/[a-z]$//\' )" && \
+            curl -fLO "https://www.openssl.org/source/openssl-$OPENSSL_VERSION.tar.gz" || \
+                curl -fLO "https://www.openssl.org/source/old/$short_version/openssl-$OPENSSL_VERSION.tar.gz" && \
+            tar xvzf "openssl-$OPENSSL_VERSION.tar.gz" && cd "openssl-$OPENSSL_VERSION" && \
+            env CC=musl-gcc ./Configure no-shared no-zlib -fPIC --prefix=/usr/local/musl -DOPENSSL_NO_SECURE_MEMORY linux-x86_64 && \
+            env C_INCLUDE_PATH=/usr/local/musl/include/ make depend && \
+            env C_INCLUDE_PATH=/usr/local/musl/include/ make && \
+            sudo make install && \
+            sudo rm /usr/local/musl/include/linux /usr/local/musl/include/asm /usr/local/musl/include/asm-generic && \
+            rm -r /tmp/*
+            export OPENSSL_DIR=/usr/local/musl/
+            export OPENSSL_INCLUDE_DIR=/usr/local/musl/include/
+            export DEP_OPENSSL_INCLUDE=/usr/local/musl/include/
+            export OPENSSL_LIB_DIR=/usr/local/musl/lib/
+            export OPENSSL_STATIC=1
+            export PQ_LIB_STATIC_X86_64_UNKNOWN_LINUX_MUSL=1
+            export PG_CONFIG_X86_64_UNKNOWN_LINUX_GNU=/usr/bin/pg_config
+            export PKG_CONFIG_ALLOW_CROSS=true
+            export PKG_CONFIG_ALL_STATIC=true
+            export LIBZ_SYS_STATIC=1
+            export TARGET=musl
         '
         ;;
 
@@ -225,8 +251,8 @@ case "$PACKAGE_OS" in
                 MAKE_FLAGS="$MAKE_FLAGS 'STRIP_COMMAND=strip'"
                 ;;
             arm32v7)
-                MAKE_FLAGS="'CARGOFLAGS=$CARGOFLAGS --target armv7-unknown-linux-gnueabihf'"
-                MAKE_FLAGS="$MAKE_FLAGS 'TARGET=target/armv7-unknown-linux-gnueabihf/release'"
+                MAKE_FLAGS="'CARGOFLAGS=$CARGOFLAGS --target armv7-unknown-linux-musleabihf'"
+                MAKE_FLAGS="$MAKE_FLAGS 'TARGET=target/armv7-unknown-linux-musleabihf/release'"
                 MAKE_FLAGS="$MAKE_FLAGS 'STRIP_COMMAND=/usr/arm-linux-gnueabihf/bin/strip'"
                 ;;
             aarch64)
